+++ casanovo/denovo/model_runner.py
@@ -36,12 +36,15 @@ class ModelRunner:
     model_filename : str, optional
         The model filename is required for eval and de novo modes,
         but not for training a model from scratch.
+    output_rootname : str, optional
+        The rootname for all output files (e.g. checkpoints or results)
     """
 
     def __init__(
         self,
         config: Config,
         model_filename: Optional[str] = None,
+        output_rootname: Optional[str] = None,
     ) -> None:
         """Initialize a ModelRunner"""
         self.config = config
@@ -54,6 +57,10 @@ class ModelRunner:
         self.loaders = None
         self.writer = None
 
+        best_filename = "best"
+        if output_rootname is not None:
+            best_filename = f"{output_rootname}.{best_filename}"
+
         # Configure checkpoints.
         self.callbacks = [
             ModelCheckpoint(
@@ -63,7 +70,7 @@ class ModelRunner:
             ModelCheckpoint(
                 dirpath=config.model_save_folder_path,
                 monitor="valid_CELoss",
-                filename="best",
+                filename=best_filename,
             ),
         ]